scripts/requirements.txt
supabase>=2.3.4
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.2.0
lxml>=5.1.0
python-dotenv>=1.0.1
